Register.cs
We added the registration page. New users need to be registered for them to be able to use
to use the system functionalities.
At a later stage we want to create an admin authorization funtionality  to control all the registration on the 
system. in order to control access to the system