OrderObj
This class object. gets all the information about an order. 
we also made a list that will take in all the number of components per order
as well as the amount or price of each components.
Once the order is done we get the status of the order
in order for the clerk to clear the order or not